+++ server/app/views/clients/_client_details.html.erb
@@ -53,7 +53,7 @@
         </div>
       </div>
       <% if current_user.super_user %>
-        <%= link_to "Print Label", client_label_path(@client.unix_user, format: :pdf), { class: "btn btn-light me-3", target: "_blank" } %>
+        <%= link_to "Print Label", pdf_label_client_path(@client.unix_user, format: :pdf), { class: "btn btn-light me-3", target: "_blank" } %>
       <% end %>
       <%= link_to "Edit Pod", edit_client_path(@client.unix_user), { class: "btn btn-light-primary me-3", data: { "turbo-frame" => "edit_client_modal" } } %>
       <a href="<%= client_measurements_path(@client.unix_user) %>" class="btn btn-light-primary me-3">View Measurements</a>
